+++ lib/api/response_models/user_response.dart
@@ -1,0 +1,8 @@
+import 'package:client_common/api/response_models/api_response.dart';
+import 'package:client_common/api/response_models/user.dart';
+
+class UserResponse extends ApiResponse {
+  User user;
+
+  UserResponse.fromJson(Map<String, dynamic> json, Map<String, String> headers) : user = User.fromJson(json);
+}
