+++ lib/navigator/guard.dart
@@ -8,7 +8,6 @@ import 'package:client_common/models/user_application_model.dart';
 import 'package:client_common/navigator/common_navigator.dart';
 import 'package:client_common/oauth/oauth_model.dart';
 import 'package:flutter/material.dart';
-import 'package:oauth2_client/access_token_response.dart';
 import 'package:provider/provider.dart';
 
 /// This class defines guards that are used to stop the user from accessing certain pages.
@@ -45,7 +44,7 @@ class Guard {
 
   static Future<bool> Function(BuildContext) _isAuthenticated(bool mustBeAuthenticated) {
     return (BuildContext context) async {
-      AuthModel authModel = context.read<AuthModel>();
+      // AuthModel authModel = context.read<AuthModel>();
       // if (!authModel.isAuthenticated() && authModel.refreshStatus.isNone()) {
       //   try {
       //     await authModel.refresh();
@@ -56,18 +55,13 @@ class Guard {
 
       // return authModel.isAuthenticated() == mustBeAuthenticated;
 
-      print("CHECK AUTHENTICATED GUARD");
-
       OAuthModel oauthModel = context.read<OAuthModel>();
-      if (!authModel.isAuthenticated()) {
-        print("NOT AUTHENTICATED");
-        AccessTokenResponse? response = await oauthModel.authenticate();
-        print("ACCESS TOKEN FROM OAUTH");
-        print(response);
-        return true;
-      }
+      // TODO: Is there a refresh feature on oauth2 ?
+      // if (!authModel.isAuthenticated()) {
+      //   return authModel.isAuthenticated() == mustBeAuthenticated;
+      // }
 
-      return false;
+      return (oauthModel.accessToken != null) == mustBeAuthenticated;
     };
   }
